src/cloudFunctions/acquireData/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage
google-cloud-secret-manager
pyowm
pytz
datetime
